added initial tests for using OpenCL Kernels with a HelloOCL Kernel simple parallelization of CPU code via openmp (currently required) switched to single-precision computation for pointclouds

+++ include/artekmed/PointCloudProcessing.h
@@ -15,18 +15,18 @@
 void buildColoredPointCloud(
         const cv::Mat& depth_img_rect,
         const cv::Mat& color_img_rect,
-        const Eigen::Matrix3d& intr_rect_ir,
+        const Eigen::Matrix3f& intr_rect_ir,
         open3d::PointCloud& cloud,
-        double depth_scale_factor=1.0);
+        float depth_scale_factor=1.0);
 
 void buildColoredPointCloud(
         const cv::Mat& depth_img_rect,
         const cv::Mat& color_img_rect,
-        const Eigen::Matrix3d& intr_rect_ir,
-        const Eigen::Matrix3d& intr_rect_rgb,
-        const Eigen::Matrix4d& depth2color_tf,
+        const Eigen::Matrix3f& intr_rect_ir,
+        const Eigen::Matrix3f& intr_rect_rgb,
+        const Eigen::Matrix4f& depth2color_tf,
         open3d::PointCloud& cloud,
-        double depth_scale_factor=1.0);
+        float depth_scale_factor=1.0);
 
 
 #endif //ARTEKMED_P1_POINTCLOUDPROCESSING_H
